feat: update services

+++ api/src/services/usersService.js
@@ -1,4 +1,4 @@
-import { and, eq } from 'drizzle-orm';
+import { eq } from 'drizzle-orm';
 import users from '../models/users.js';
 import db from '../utils/dbManager.js';
 import { hashSecret } from '../utils/secretHashing.js';
@@ -30,17 +30,12 @@ export async function getUserByEmail(email) {
 }
 
 export async function deleteUser(req) {
-  const { email, password } = req.body;
-  validateEmailAndPassword(email, password);
-  const hashCheck = await hashSecret(password);
+  const userId = req.user.id;
   const removed = await db.delete(users)
-    .where(and(eq(users.email, email), users.hashedPassword === hashCheck))
+    .where(eq(users.id, userId))
     .returning();
   if (removed.length < 1) {
-    const userExists = await db.select()
-      .from(users)
-      .where(eq(users.email, email));
-    if (userExists.length < 1) {
+    if (removed.length < 1) {
       throw new Error('No user found with the provided email');
     }
     throw new Error('The password is incorrect');
@@ -49,9 +44,9 @@ export async function deleteUser(req) {
 
 export async function updateUser(req) {
   const {
-    email, password, newEmail, newPassword,
+    newEmail, newPassword,
   } = req.body;
-  validateEmailAndPassword(email, password);
+  const userId = req.user.id;
   const newObj = {};
   if (!newEmail && !newPassword) {
     throw new Error('No parameters supplied');
@@ -60,16 +55,15 @@ export async function updateUser(req) {
     newObj.email = newEmail;
   }
   if (newPassword) {
-    newObj.hashedPassword = hashSecret(password);
+    newObj.hashedPassword = await hashSecret(newPassword);
   }
   newObj.updated = new Date();
-  const hashCheck = hashSecret(password);
   const result = await db.update(users)
     .set(newObj)
-    .where(and(eq(email, users.email), users.hashedPassword === hashCheck))
+    .where(eq(userId, users.id))
     .returning({ id: users.id, email: users.email });
   if (result.length < 1) {
-    throw new Error('Email/password incorrect');
+    throw new Error('Invalid API token');
   }
   return result[0];
 }
